Refactor initial syncer classes
- add post method to ApiManager and remove unused methods
- remove HttpRequester class use ApiManager instead
- move BCoinApi class to a separate file
